rename cmd/cli to cmd/infrakit
