+++ cmd/infrakit/template/template.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/docker/infrakit/cmd/cli/base"
+	"github.com/docker/infrakit/cmd/infrakit/base"
 	"github.com/docker/infrakit/pkg/discovery"
 	logutil "github.com/docker/infrakit/pkg/log"
 	"github.com/ghodss/yaml"
